pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "audiobiblio"
version = "0.2.0"
description = "Audiobook crawler/downloader + tag fixer for mujrozhlas.cz"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "haxny" }]
license = "MIT"

dependencies = [
  "SQLAlchemy>=2.0",
  "alembic>=1.13",
  "click>=8.1",
  "structlog>=24.1",
  "pydantic>=2.7",
  "typer>=0.12",
  "yt-dlp>=2024.7.7",                   # unpinned as you wanted
  "mutagen>=1.47",
  "requests>=2.32",
  "Pillow>=10.0",
  "rich>=13.7",
  "beautifulsoup4>=4.12",
  "platformdirs>=4.2",
  "PyYAML>=6.0",
  "APScheduler>=3.10",
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "jinja2>=3.1",
  "python-multipart>=0.0.9",
  "sse-starlette>=2.0",
]

[project.optional-dependencies]
fs = ["xxhash>=3.4"]
downloader = ["Pillow", "yt-dlp", "requests", "mutagen"]
tags = ["mutagen", "rich", "structlog", "PyYAML"]

[project.scripts]
audiobiblio = "audiobiblio.cli:app"
audioloader = "audiobiblio.audioloader:main"
tag-fixer   = "audiobiblio.tag_fixer:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["audiobiblio*"]